Shared memory (from SMPI)
We didn't take everything from SMPI yet (no automatic sharing for
example, we only have the manual macros). It seems to work well on
all our test suite, but it's hard to use in MPI, because some
buffers contain import meta-data, so sharing memory in MPI should
be done preferably at the application level. If we want to do it in
MPI directly it will require a deep analysis of what it's trying to
do with its memory