Fix some issues in PCI model
We didn't have a lot of problems until now because we did very few
intra-node communications, but our PCI model was buggy because we
were only synchronizing on BXI transfers and not PCI ones. So this
fixes this issue, using callbacks at the end of s4u::Comms (it's a
shame this mechanism isn't well documented because it's really
useful).

This also allows to remove our horrible "dma_actor"s to LOG things
at the end of transfers