Sync Initiator on PCI comm instead of PCI and BXI
This isn't right either but it feels less wrong: by having the Initiator
wait for both the PCI and BXI comm before processing a new message
we basically kill any possibility of having several messages inflight
(I mean, in the general case we can have at most exactly N messages
inflight if we have N initiators per NIC, which is a problem since N
doesn't really represent anything that's precisely measurable in the
real hardware)

The problem now is that waiting for the PCI comm only works really
well when there is no congestion on the BXI network, but if there
is the initiators are still going to inject messages at PCI speed
even though in the real world the lack of flit-level credits would
forbid them to do so, so that's a problem, but I don't think I can
fix it without making a detailed switch model (and thus
reimplementing all the routing algorithm, so all the AFM basically).

The problem is that we can sync on the BXI comm completely finishing
(i.e. arriving at the target NIC), but we don't have access to the
event "all the packets are on the cable, even though they might not
be at target yet", or even the event "the message completely reached
the first switch". This is probably fundamentally impossible right
now because of the wormhole routing at the packet level, and the way
that SimGrid handles comms, so we'll see what we do in the future
about that