Allow MPI middleware to be disabled in CMake
Toggling the middleware should now be possible simply by changing
the "BUILD_MPI_MIDDLEWARE" CMake variable (or environment variable
if using the rebuild.sh script).

For now this means we'll disable shared mallocs / free if the middleware
is not enabled (it seems stupid but this is linked with all the sample /
bench functions being private in SMPI and not public in SimGrid, which
is very sad).

It also means that we bring back our copy / pasted bench model if the
middleware is disabled (same issue, much sadness).