Allow polling functions to have negative values
I don't know this worked in the first place, as OMPI_REQUEST_PML is
zero, so using zero as a default value was an error.

Also add an extern int value to indicate whether polling should be
allowed or not. This should probably not be here, as the code doesn't
use it at all, it's just a convenient place to store wether vader is in
use or not (from OpenMPI)